requirements.txt
nltk
numpy
tqdm
pillow
numpy
pandas
torch
transformers>=4.0.0
google-genai
